Introduce shortcode img
This way the img is centered

I'm really not too fond of this as it renders classical markdown editors
useless. Then again if the img is not a relative path, they won't be
able to find it anyway.

It would be interesting to see what are the commonly used tools for
writing posts.