Move body's padding out of baseof
I'm not sure if this is a good idea, but I need it as for the index page
I don't want there to be any padding.

All of this CSS is getting a bit un-maintainable only because of all the
code duplication. I definitely see the need of using @apply from
tailwind in the future to avoid so much duplication.